+++ CHANGELOG.md
@@ -9,7 +9,12 @@ CHANGELOG
 		uid:<string>
 		prototype:<string>
 		properties:<json schema Properties>
-		geometry:j<son schema GeoJson>
+		geometry:<json schema GeoJson>
+
+- analogo discorso per  PUT `/feature/{uid}`:
+
+		properties:<json schema Properties>
+		geometry:<json schema GeoJson>
 
 - sistemata la risposta di GET `/feature`, modificando il livello di "geometry"
 - implementata api PUT `/features/{uid}`. Accetta il paramentro `properties` con uno schema analogo al parmetro `feature` di POST `/features` con queste differenze: **nello schema della PUT, `uid` e `prototype` NON vengono accettati perchè sono campi chiave della feature e non possono essere modificati**. lo `uid` della feature deve essere specificato come url e non come parametro.
@@ -86,7 +91,7 @@ CHANGELOG
     nota 4: ___"state"___ viene mantenuto ma per ora è inutilizzato
 
         {
-            features": [
+            "features": [
                 {
                     "type": "Feature",
                     "properties": {